작성자 : 이규호 내용 : 컨트롤러/링크 수정
