+++ CoDream/src/main/java/com/code/dream/HomeController.java
@@ -25,8 +25,6 @@ import com.code.dream.security.UserSecurityDto;
 @Controller
 public class HomeController {
 	
-	@Autowired
-	private IRegteacherService service;
 	
 	private static final Logger logger = LoggerFactory.getLogger(HomeController.class);
 	
@@ -47,43 +45,5 @@ public class HomeController {
 		return "home";
 	}
 	
-	@RequestMapping(value = "/regteacherForm", method = RequestMethod.GET)
-	public String regteacherForm(Model model, Authentication authentication) {
-		UserSecurityDto usDto = (UserSecurityDto) authentication.getPrincipal();
-		RegisterDto dto = usDto.getDto();
-		dto.setPassword(null);
-		model.addAttribute("dto", dto);
-		return "admin/regteacherForm";
-	}
 	
-	@RequestMapping(value = "/regteacher", method = RequestMethod.POST)
-	public String regteacher(RegteacherDto dto, Authentication authentication) {
-		UserSecurityDto usDto = (UserSecurityDto) authentication.getPrincipal();
-		RegisterDto rdto = usDto.getDto();
-		dto.setUserid(rdto.getId());
-		service.insertRegteacher(dto);
-		return "redirect:/admin/regteacherList";
-	}
-	
-	@RequestMapping(value = "/admin/regteacherList", method = RequestMethod.GET)
-	public String regteacherList(Model model) {
-		List<RegteacherDto> list = service.selectRegteacher();
-		model.addAttribute("list", list);
-		return "/admin/regteacherList";
-	}
-	
-	@RequestMapping(value = "/regteacherDetail", method = RequestMethod.GET)
-	public String regteacherDetail(Model model, String te_seq) {
-		RegteacherDto dto = service.detailRegteacher(te_seq);
-		List<AttachFileDto> list = service.selectFiles(te_seq);
-		model.addAttribute("dto", dto);
-		model.addAttribute("list", list);
-		return "admin/regteacherDetail";
-	}
-	
-	@RequestMapping(value = "/regteacherModify", method = RequestMethod.GET)
-	public String regteacherModify(String te_seq, String te_admit) {
-		service.modifyRegteacher(te_admit, te_seq);
-		return "redirect:/admin/regteacherList";
-	}
 }
